rm IOConfig and replace with VarTensor methods
