+++ src/tensor/val.rs
@@ -4,7 +4,7 @@ use super::*;
 /// Halo2's `Value<F>`, `Value<Assigned<F>>`, `AssignedCell<Assigned<F>, F>`.
 /// This enum is generally used to assign values to variables / advices already configured in a Halo2 circuit (usually represented as a `VarTensor`).
 /// For instance a `ValTensor` can represent pre-trained neural network weights; or a known input to a network.
-/// The `nn.io` module provides helper structs and methods to do this assignment.
+/// The `VarTensor` enum provides helper methods to do this assignment.
 #[derive(Debug, Clone)]
 pub enum ValTensor<F: FieldExt + TensorType> {
     Value {
@@ -49,6 +49,7 @@ impl<F: FieldExt + TensorType> From<Tensor<AssignedCell<Assigned<F>, F>>> for Va
 }
 
 impl<F: FieldExt + TensorType> ValTensor<F> {
+    /// Calls `get_slice` on the inner tensor.
     pub fn get_slice(&self, indices: &[Range<usize>]) -> ValTensor<F> {
         match self {
             ValTensor::Value { inner: v, dims: _ } => {
@@ -75,6 +76,7 @@ impl<F: FieldExt + TensorType> ValTensor<F> {
         }
     }
 
+    /// Sets the `ValTensor`'s shape.
     pub fn reshape(&mut self, new_dims: &[usize]) {
         match self {
             ValTensor::Value { inner: _, dims: d } => {
@@ -92,6 +94,7 @@ impl<F: FieldExt + TensorType> ValTensor<F> {
         }
     }
 
+    /// Calls `flatten` on the inner tensor.
     pub fn flatten(&mut self) {
         match self {
             ValTensor::Value { inner: v, dims: d } => {
@@ -109,6 +112,7 @@ impl<F: FieldExt + TensorType> ValTensor<F> {
         }
     }
 
+    /// Returns the `dims` attribute of the `ValTensor`.
     pub fn dims(&self) -> &[usize] {
         match self {
             ValTensor::Value { inner: _, dims: d } => d,
